refactor: apis/event-board 폴더구조 변경
