+++ src/services/apis/event-board/event-board.type.ts
